+++ Tests/SwiftStdlibTests/DictionaryExtensionsTests.swift
@@ -78,6 +78,16 @@ final class DictionaryExtensionsTests: XCTestCase {
         XCTAssertEqual(dict, ["testkey": "value"])
     }
 
+    func testSubscriptKeypath() {
+        var json = ["key": ["key1": ["key2": "value"]]]
+
+        XCTAssertEqual(json[path: []] as? String, nil)
+        XCTAssertEqual(json[path: ["key", "key1"]] as? [String: String], ["key2": "value"])
+        XCTAssertEqual(json[path: ["key", "key1", "key2"]] as? String, "value")
+        json[path: ["key", "key1", "key2"]] = "newValue"
+        XCTAssertEqual(json[path: ["key", "key1", "key2"]] as? String, "newValue")
+    }
+
     func testOperatorPlus() {
         let dict: [String: String] = ["key1": "value1"]
         let dict2: [String: String] = ["key2": "value2"]